All the logic for the search and showing results

+++ app/src/main/java/com/example/weatherapp/model/geocodingmodel/Rank.kt
@@ -1,7 +1,7 @@
 package com.example.weatherapp.model.geocodingmodel
 
 data class Rank(
-    val confidence: Int,
+    val confidence: Double,
     val confidence_city_level: Int,
     val importance: Double,
     val match_type: String
